requirements.txt
python-decouple==3.8
Pyrogram==2.0.106
peewee==3.17.6
prometheus-client==0.19.0